Move server bootstrap behind app package
Move authgate runtime wiring out of cmd/authgate/main.go into internal/app so the command package stays a thin entrypoint. Route wiring tests move with the unexported route helpers, and the architecture package map is updated in the same change.

Constraint: The repository architecture keeps built-in handlers/services flat while zitadel/oidc storage remains in internal/storage; this refactor only moves process bootstrap wiring.
Rejected: Splitting storage by subpackage | zitadel/oidc requires one Storage type to satisfy op.Storage and DeviceAuthorizationStorage.
Confidence: high
Scope-risk: moderate
Directive: Keep cmd/authgate/main.go free of service, route, metrics, cleanup, and provider wiring.
Tested: git diff --check; go test ./...; go vet ./...
Not-tested: Live server startup against a production database.
