+++ internal/app/routes_test.go
@@ -1,4 +1,4 @@
-package main
+package app
 
 import (
 	"net/http"
